+++ src/test/java/com/github/grishberg/tests/TestUtils.java
@@ -1,0 +1,27 @@
+package com.github.grishberg.tests;
+
+import java.io.BufferedReader;
+import java.io.FileInputStream;
+import java.io.InputStreamReader;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Created by grishberg on 28.04.18.
+ */
+public class TestUtils {
+    public static List<String> readFile(String fileName) throws Exception{
+
+        ArrayList<String> lines = new ArrayList<>();
+
+        try (BufferedReader br = new BufferedReader(
+                new InputStreamReader(
+                        new FileInputStream(fileName), "UTF-8"))) {
+            String line;
+            while ((line = br.readLine()) != null) {
+                lines.add(line);
+            }
+        }
+        return lines;
+    }
+}